Port 'Don't generate code for document.all check in typeofisandbranch' to M40
Original description

For compatibility with old websites that call
`typeof document.all == "undefined"`, V8 generates code
 to check if the typeof operand is document.all, in all
cases (typeof document.all must give "undefined", not
"object").

However, in the benchmark typeof check it can be seen
that the expression cannot be document.all, so generation
of this code can be avoided.

Gives roughly a 4-5% improvement to the Array_Blur score
on Note 3 Android 4.4.

Bug: http://107.108.218.239/bugzilla/show_bug.cgi?id=1533
Reviewed by:  Heejin Chung, Kangil Han, Viatcheslav Ostapenko

Change-Id: I9f3353eba583ad1fa1c38797140e81a60463ff07
Signed-off-by: Petka Antonov <p.antonov@partner.samsung.com>