Integrate REGEXP_WITH_FUNCTION.
This patch will improve Sunspider 1.0.2 by ~4% and Kraken 1.1 by ~3%.
The purpose of this patch is to cache strings of regular expression
with function for a large script.

This cache will not be enabled unless the input script is very long, we
don’t want to add overhead checking in short script case.

An extra string pointer is saved in the cache, 4 bytes for ARM and 8
bytes for ARM64 platform.

The cache is cleared by global GC.

	    Before   After
Sunspider   802      761
Kraken      5144     5184
V8          6523     6538
Octane      6413     6302

Bug: http://web.sec.samsung.net/bugzilla/show_bug.cgi?id=13898

Original Patch by: a.bangash <a.bangash@samsung.com>
Signed-off-by: chevalam.s <chevalam.s@samsung.com>
Reviewed by: DaeHyun Ko, Siba Samal

Change-Id: I44dad4f5ef5db14c1666a068a0cc31b70cd254a5
Signed-off-by: chevalam.s <chevalam.s@samsung.com>